// src/app/services/generated/model/cameraStream.ts
/**
 * Read management system Frontend API
 * Read management system Frontend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CameraStreamProtocol } from './cameraStreamProtocol';
import { Voivodeship } from './voivodeship';
import { Localization } from './localization';


export interface CameraStream { 
    id?: number;
    name: string;
    localization: Localization;
    voivodeship: Voivodeship;
    url: string;
    protocol: CameraStreamProtocol;
}